refactor(tools): seperate uuid functions from tools.utils
