+++ kong-3.6.0-0.rockspec
@@ -169,6 +169,7 @@ build = {
     ["kong.tools.table"] = "kong/tools/table.lua",
     ["kong.tools.sha256"] = "kong/tools/sha256.lua",
     ["kong.tools.yield"] = "kong/tools/yield.lua",
+    ["kong.tools.uuid"] = "kong/tools/uuid.lua",
 
     ["kong.runloop.handler"] = "kong/runloop/handler.lua",
     ["kong.runloop.events"] = "kong/runloop/events.lua",
